+++ sdk/confidentialledger/Azure.Security.ConfidentialLedger/src/ConfidentialLedgerIdentityServiceClient.cs
@@ -33,8 +33,8 @@ namespace Azure.Security.ConfidentialLedger
             options ??= new ConfidentialLedgerClientOptions();
             _clientDiagnostics = new ClientDiagnostics(options);
             _pipeline = HttpPipelineBuilder.Build(options, new HttpPipelinePolicy[] { new LowLevelCallbackPolicy() }, Array.Empty<HttpPipelinePolicy>(), new ResponseClassifier());
-            this.identityServiceUri = identityServiceUri;
-            apiVersion = options.Version;
+            _identityServiceUri = identityServiceUri;
+            _apiVersion = options.Version;
         }
 
         /// <summary>