+++ sdk/storage/Azure.Storage.Blobs/src/Models/BlobDownloadOptions.cs
@@ -8,19 +8,29 @@ using Tags = System.Collections.Generic.IDictionary<string, string>;
 namespace Azure.Storage.Blobs.Models
 {
     /// <summary>
-    /// Optional paratmers for dowloading to a Blob.
+    /// Optional parameters for downloading a range of a blob.
     /// </summary>
     public class BlobDownloadOptions
     {
         /// <summary>
-        /// Optional <see cref="IProgress{Long}"/> to provide
-        /// progress updates about data transfers.
+        /// If provided, only download the bytes of the blob in the specified
+        /// range.  If not provided, download the entire blob.
         /// </summary>
-        public IProgress<long> ProgressHandler { get; set; }
+        public HttpRange Range { get; set; }
+
         /// <summary>
-        /// Optional <see cref="StorageTransferOptions"/> to configure
-        /// parallel transfer behavior.
+        /// Optional <see cref="BlobRequestConditions"/> to add conditions on
+        /// downloading this blob.
         /// </summary>
-        public StorageTransferOptions TransferOptions { get; set; }
+        public BlobRequestConditions Conditions { get; set; }
+
+        /// <summary>
+        /// Optional transactional hashing options.
+        /// Range must be provided explicitly stating a range withing Azure
+        /// Storage size limits for requesting a transactional hash. See the
+        /// <a href="https://docs.microsoft.com/en-us/rest/api/storageservices/get-blob">
+        /// REST documentation</a> for range limitation details.
+        /// </summary>
+        public DownloadTransactionalHashingOptions TransactionalHashingOptions { get; set; }
     }
 }