+++ sdk/webpubsub/Microsoft.Azure.WebPubSub.Common/src/Internal/WebPubSubClientCertificateJsonConverter.cs
@@ -1,0 +1,28 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+using System;
+using System.Text.Json;
+using System.Text.Json.Serialization;
+
+namespace Microsoft.Azure.WebPubSub.Common
+{
+    internal class WebPubSubClientCertificateJsonConverter : JsonConverter<WebPubSubClientCertificate>
+    {
+        public override WebPubSubClientCertificate Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
+        {
+            var element = JsonDocument.ParseValue(ref reader).RootElement;
+
+            return new WebPubSubClientCertificate(
+                element.ReadString(WebPubSubClientCertificate.ThumbprintProperty));
+        }
+
+        public override void Write(Utf8JsonWriter writer, WebPubSubClientCertificate value, JsonSerializerOptions options)
+        {
+            writer.WriteStartObject();
+            writer.WritePropertyName(WebPubSubClientCertificate.ThumbprintProperty);
+            JsonSerializer.Serialize(writer, value.Thumbprint, options);
+            writer.WriteEndObject();
+        }
+    }
+}