+++ sdk/storage/Azure.Storage.Common/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Release History
 
 ## 12.10.0-beta.1 (Unreleased)
+- Added support for SDK-calculated transactional hash checksums on data transfer.
 - This release contains bug fixes to improve quality.
 
 ## 12.9.0 (2021-09-08)